small change on vehicle, all int into double

+++ Ta todo la dentro/src/logic/Vehicle.java
@@ -3,12 +3,12 @@ package logic;
 public class Vehicle {
 	
 	
-	private int currentFuel;
-	private int maxCapacity;
-	private int consumption;
+	private double currentFuel;
+	private double maxCapacity;
+	private double consumption;
 
 	
-	public Vehicle(int start, int max){
+	public Vehicle(double start, double max){
 		if(start > max){
 			currentFuel = max;
 		} else {
@@ -18,7 +18,7 @@ public class Vehicle {
 		consumption = 1;
 	}
 	
-	public Vehicle(int start, int max, int cons){
+	public Vehicle(double start, double max, double cons){
 		if(start > max){
 			currentFuel = max;
 		} else {
@@ -28,15 +28,15 @@ public class Vehicle {
 		consumption = cons;
 	}
 
-	public int getCurrentFuel(){
+	public double getCurrentFuel(){
 		return currentFuel;
 	}
 	
-	public int getMaxCapacity(){
+	public double getMaxCapacity(){
 		return maxCapacity;
 	}
 	
-	public int getConsumption() {
+	public double getConsumption() {
 		return consumption;
 	}
 	
@@ -48,7 +48,7 @@ public class Vehicle {
 		}
 	}
 	
-	public void travel(int distance) {
+	public void travel(double distance) {
 		currentFuel -= distance*consumption;
 	}
 	
@@ -56,11 +56,11 @@ public class Vehicle {
 		currentFuel = maxCapacity;
 	}
 	
-	public void setMaxCapacity(int cap){
+	public void setMaxCapacity(double cap){
 		maxCapacity = cap;
 	}
 
-	public void setConsumption(int cons) {
+	public void setConsumption(double cons) {
 		consumption = cons;
 	}
